requirements.txt
flask==2.1.0
beautifulsoup4==4.11.0
requests == 2.28.1
Flask-SQLAlchemy==2.5.1
python-dotenv==0.20.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
